Put combine data actions into view data window
