--- combine_data.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Combine Data</title>
-    <link rel="stylesheet" href="combine_data.css">
-</head>
-
-<body>
-    <header>
-        <h3>Combine Files</h3>
-    </header>
-    <main>
-        <div id="files-container">
-            <label for="file-input" id="files">Add Files</label>
-            <input id="file-input" type="file" accept=".json,application/json" multiple></input>
-            <div id="list-container">
-                Files
-                <!-- This list will be populated by javascript -->
-                <ul id="file-list"></ul>
-            </div>
-        </div>
-        <p>&rarr;</p>
-        <div id="input-container">
-            <input type="text" id="output-file" placeholder="Ouput File Name">
-        </div>
-        <p>&rarr;</p>
-        <div id="output-container">
-            <button id="csv-button">Make CSV File</button>
-        </div>
-    </main>
-    <footer>
-    </footer>
-    <script src="combine_data.js"></script>
-</body>
-</html>
