+++ index.html
@@ -24,7 +24,6 @@
         <button id="submit">Submit</button>
         <button id="reset">Reset</button>
         <button id="view">View Data</button>
-        <button id="combine-data">Combine Data</button>
     </div>
     <div class="row">
         <h1>AUTO</h1>
